Get rid of warning in game project
warning C4996: 'std::_Transform1': Function call with parameters that
may be unsafe - this call relies on the caller to check that the passed
values are correct.

This is a warnings that Visual Studio spits out if you use any of the
Standart C++ Library functions that Microsoft has provided secure
replacements for (such as strncpy). This isn’t necessarily a bad idea,
but their safe functions are not part of the standard and are thus
non-portable. I disabled the warning.