Fixing bug with Warrior Execute
Warrior execute is actually working but not notifying the
client as it should. Once the execute is missed/dodged/parried
on default spellId's sent by the client, the spell miss message
isn't meaningfull for the client. In fact, it only reacts on
spell 20647 to display the result (damage or miss result).

Secondary, fixing the spell 20647 to avoid that it can actually miss
while the prior spell passed succesfully. This is asumption since the
DBC are reporting that the EquipItemID is -1 for the spell 20647
(making MeleeHitResult calculation false).

Signed-off-by: Talendrys <cedric_lesi@hotmail.com>