[Docker] Use supervisor instead of launch script
This commit removes the launch script that checked for the existence of
data containers and extracted data. It expects the user to mount the
appropriate folders into the correct locations. Three volumes are
expected:

* /opt/mangos/data - (required) must contain DBCs, maps etc.
* /opt/mangos/etc - (required) must contain configuration files
* /opt/mangos/logs - (optional) stdout/stderr and all logs will be written
  to this folder. If you want to collect them off the container, use a volume.

In addition, instead of just forking `realmd` into the background and
running `mangosd` as the main container process, the Dockerfile now
installs `supervisord` and supervises both processes using it.
The configuration for this is in `docker-supervisord.conf`