Blacksmithing and Leatherworking specialisation scripts removed
For the way these work in Vanilla, the scripts are not required.

Player characters have to unlearn the entire profession, in order to
gain a different specialisation.

NOTE: STILL TO DO - make sure the specialisation quests are removed when
the profession is unlearnt, so that the player character can acquire a
specialisation again.